+++ R/coreSectionDepth.R
@@ -188,9 +188,42 @@ multi_coreSection_to_dblf <- function(corename,cm){
     cm <- unlist(cm)
   }
 
-
   return(purrr::map2_dfr(corename,cm,.f = coreSection_to_dblf))
 }
+
+
+
+
+
+#' Convert HSI depths for multiple core sections
+#'
+#' @param corename a vector of corenames
+#' @param cm a vector of corresponding depths that matches the length of corenames
+#'
+#' @importFrom purrr map2_dfr
+#' @return a tibble
+#' @export
+#'
+#' @examples
+#'
+#' multi_hsi_to_dblf(c("L380_DUNCA3_LC4U_1","L380_DUNCA3_LC4U_2"),c(20,10))
+#'
+multi_hsi_to_dblf <- function(corename,cm){
+
+
+  if(is.list(corename)){
+    corename <- unlist(corename)
+  }
+
+
+  if(is.list(cm)){
+    cm <- unlist(cm)
+  }
+
+  return(purrr::map2_dfr(corename,cm,.f = hsi_to_dblf))
+}
+
+
 
 
 #' List known core names
@@ -225,3 +258,77 @@ findCoreSectionName <- function(corename_guess){
   }
   return(finalKey$`Section Name`[index])
 }
+
+
+#' Convert depths from hyperspectral imaging
+#'
+#' @param corename a vector of corenames
+#' @param cm a vector of corresponding depths that matches the length of corenames
+#'
+#' @return
+#' @export
+#'
+#' @examples
+#'
+#' hsi_to_dblf("L380_DUNCA3_LC4U_2",0:30)
+#'
+hsi_to_dblf <- function(corename,cm){
+
+  if(all(is.na(cm))){
+    return(NA)
+  }
+
+  if(!is.character(corename)){
+    stop("corename should be of class character")
+  }
+
+  corename <- unique(corename)
+  if(length(corename) > 1){
+    stop("more than one corename was entered. To calculate depths from multiple sections, use multi_hsi_to_dblf")
+  }
+
+  #find the relevant core section row
+  section <- dplyr::filter(finalKey,tolower(corename) == tolower(`Section Name`))
+
+  if(nrow(section) == 0){
+    stop(glue::glue("Couldn't find a core section named {corename}. Search for corenames with `findCoreSectionName()`, or for a list of known core sections run `listCoreSectionNames()`"))
+  }
+
+  if(nrow(section) > 1){
+    stop("Multiple core section matches. This shouldn't happen")
+  }
+
+  # Get key metadata
+  secTopDblf <- section$secTopDblf[1]
+  secRoiTop <- section$roiTop[1]
+  secRoiBot <- section$roiBot[1]
+
+  #adjust to bottom depth if within 1 cm.
+  if(any(between(secRoiBot - cm,-1,0))){
+    tc <- which(between(secRoiBot - cm,-1,0))
+    cm[tc] <- secRoiBot
+  }
+
+  if(is.numeric(secRoiBot) & is.numeric(secRoiTop)){
+    if(any((secRoiBot - secRoiTop) < cm)){
+      badDepth <- cm[which((secRoiBot - secRoiTop) < cm)]
+      stop(glue::glue("At least one requested depth ({badDepth[1]} cm on HSI scale, {badDepth[1] + secRoiTop} relative to the coreliner) is outside the ROI range ({secRoiTop} to {secRoiBot} cm) for core {corename}"))
+    }
+  }
+
+
+
+  if(!is.numeric(secTopDblf)){
+    stop("missing metadata for core {corename}, cannot calculate depth below lake floor")
+  }
+
+  #then calculate
+  dblf <- cm + secTopDblf
+
+  return(data.frame(dblf = dblf,
+                    topSource = section$topSource[1],
+                    botSource = section$botSource[1],
+                    coreName = corename))
+
+}
+